requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
plotly>=5.18
streamlit>=1.30
apache-airflow>=2.7
pendulum>=3.0
